// generated/src/app/api-models/ipa-affiliate-address.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class IpaAffiliateAddress {
  seqIpaAffltAddress: number;
  ipaId: string;
  ipaProvCategory: string;
  name2: string;
  addressLine1: string;
  addressLine2: string;
  city: string;
  state: string;
  zipCode: string;
  county: string;
  country: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}